requirements.txt
alembic==1.8.1
fastapi==0.79.0
uvicorn==0.18.2
celery==5.2.7
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytest==7.1.2
PyJWT==2.4.0
SQLAlchemy==1.4.39
requests==2.28.1
ipdb==0.13.9
websockets==10.3